Add interface ViewingScopeValidator
